new patch

+++ src/zensvi/download/streetview_downloader.py
@@ -650,7 +650,7 @@ class MLYDownloader(BaseDownloader):
         # drop NA values in id columns
         pid_df = pid_df.dropna(subset=['id'])
         # get unique pids (ie "id" columns) as a list
-        pids = pid_df["id"].unique().tolist()
+        pids = pid_df["id"].astype(int).unique().tolist()
         return pids
 
     def _set_dirs(self, dir_output):
